Fix dropdown rendering
Unfortunately the dropdown is set to `position: absolute` and attached
to a `position: relative` div which is nested in a div with `overflow:
hidden` set on it. That means that the dropdown can not be rendered
outside of the main content window (the white window in the middle).

The fix for this would be to unset `position: relative` on that div and
then position everything using JavaScript but that would complicate
everything immensly. Instead as a hack I decided to simply render the
dropdown above the anchor point which allows us to use the fact that
there is extra space above the track list. Thanks to that the content of
the dropdown will never even get close to the boundary of the main
content window. I guess it is now a dropup.