Fix playing indicator being highlighted erronously
After loading the frontend it was possible to click the play button.
This has no effect as the queue is empty but would change the state of
the playing indicator as it was only based on !store.paused being true.
Added a check to make sure that the play indicator is only highlighted
if the store.paused is not set and there actually is actually being
played.